+++ src/test/scala/org/apache/spark/streaming/rabbitmq/RabbitMQConsumerIT.scala
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (C) 2015 Stratio (http://stratio.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -29,7 +29,7 @@ class RabbitMQConsumerIT extends TemporalDataSuite {
 
   test("RabbitMQ Receiver should read all the records") {
 
-    val receiverStream = RabbitMQUtils.createStream[String](ssc, Map(
+    val receiverStream = RabbitMQUtils.createStream(ssc, Map(
       "hosts" -> hosts,
       "queueName" -> queueName,
       "exchangeName" -> exchangeName,
@@ -38,7 +38,7 @@ class RabbitMQConsumerIT extends TemporalDataSuite {
       "userName" -> userName,
       "password" -> password
     ))
-    val totalEvents = ssc.sparkContext.accumulator(0L, "My Accumulator")
+    val totalEvents = ssc.sparkContext.longAccumulator("My Accumulator")
 
     // Start up the receiver.
     receiverStream.start()
@@ -49,7 +49,7 @@ class RabbitMQConsumerIT extends TemporalDataSuite {
         val count = rdd.count()
         // Do something with this message
         println(s"EVENTS COUNT : \t $count")
-        totalEvents += count
+        totalEvents.add(count)
         //rdd.collect().sortBy(event => event.toInt).foreach(event => print(s"$event, "))
       } else println("RDD is empty")
       println(s"TOTAL EVENTS : \t $totalEvents")